Replacing /venues as endpoint
